refactor(mcp): split project context helpers
